+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Target functions into Label
+- Mouse Scroll functionality into ScrollToSelection
 
 ## [0.8.4] - 2026-08-31
 ### Added